requirements.txt
boto3==1.17.21
botocore==1.20.21
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
geographiclib==1.50
geopy==2.1.0
idna==2.10
jmespath==0.10.0
kiwisolver==1.3.1
matplotlib==3.3.4
mysql-connector-python==8.0.32
numpy==1.20.1
git+https://github.com/scottedkeenan/python-ogn-client-scottracks.git
Pillow==8.1.0
protobuf==3.14.0
pyparsing==2.4.7
python-dateutil==2.8.1
requests==2.25.1
s3transfer==0.3.4
scipy==1.6.0
six==1.15.0
urllib3==1.26.3

setuptools==57.0.0
pytz==2022.7.1
pandas==1.5.3
MarkupSafe==2.1.2
Jinja2==3.1.2
pika==1.2.1
folium==0.14.0
redis==5.0.1